requirements.txt
streamlit
opencv-python-headless
torch
Pillow
numpy
ultralytics
torchvision